Moved emulator to core and runner to fuzzer packages
