+++ src/main/java/gr/aueb/android/barista/fuzzer/ContextEventGenerator.java
@@ -2,7 +2,6 @@ package gr.aueb.android.barista.fuzzer;
 
 import gr.aueb.android.barista.core.context.model.*;
 import gr.aueb.android.barista.core.model.Command;
-import gr.aueb.android.barista.emulator.EmulatorManager;
 import gr.aueb.android.barista.utilities.BaristaLogger;
 
 import java.util.*;
