+++ src/main/java/gr/aueb/android/barista/fuzzer/runner/Runner.java
@@ -1,4 +1,4 @@
-package gr.aueb.android.barista.runner;
+package gr.aueb.android.barista.fuzzer.runner;
 
 public interface Runner {
