+++ src/main/java/gr/aueb/android/barista/server/HttpServerManager.java
@@ -10,8 +10,8 @@
 package gr.aueb.android.barista.server;
 
 import gr.aueb.android.barista.core.BaristaConfigurationExtension;
-import gr.aueb.android.barista.emulator.EmulatorManager;
-import gr.aueb.android.barista.emulator.helpers.TestMonitor;
+import gr.aueb.android.barista.core.emulator.EmulatorManager;
+import gr.aueb.android.barista.core.emulator.helpers.TestMonitor;
 import gr.aueb.android.barista.utilities.BaristaLogger;
 import org.glassfish.grizzly.http.server.HttpServer;
 import org.glassfish.jersey.grizzly2.httpserver.GrizzlyHttpServerFactory;
