+++ src/test/java/gr/aueb/android/barista/core/model/SvcWifiStatusTest.java
@@ -2,7 +2,7 @@ package gr.aueb.android.barista.core.model;
 
 import gr.aueb.android.barista.core.executor.CommandExecutorFactory;
 import gr.aueb.android.barista.core.executor.CommandExecutorImpl;
-import gr.aueb.android.barista.emulator.EmulatorManager;
+import gr.aueb.android.barista.core.emulator.EmulatorManager;
 import org.junit.BeforeClass;
 import org.junit.Test;
